grad_check call modified, loss refactored
grad_check uses a subset of the dataset. As grad_check was too slow because of calculating the whole dataset each time (7840 * 60k iterations for MNIST), only a subset of the dataset is given.
Loss refactored to use given y_ instead of pre-given whole dataset Y_. That way loss calculation does not depend on the dataset that is used but on the given y_.
Refactor should still be continued to decouple model from loss and dataset.